backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3
chromadb==0.5.5
sentence-transformers==3.0.1
rank-bm25==0.2.2
pypdf==5.0.1
python-docx==1.1.2
psutil==5.9.8
python-multipart==0.0.9